Refactor Aptos provider and integrate chain_traits
Replaces the typeshare-based implementation in gem_aptos with a new provider module implementing chain_traits interfaces for balances, state, tokens, transactions, and preload. Updates model types for better deserialization and compatibility. Removes typeshare files, adds new mappers and test data, and updates dependencies to use num-bigint and chain_traits. Refactors related code in gem_chain_rpc and other crates to use the new interfaces and naming conventions.

+++ crates/gem_algorand/src/provider/preload.rs
@@ -16,7 +16,7 @@ impl<C: Client> ChainPreload for AlgorandClient<C> {
             block_hash: params.genesis_hash,
             block_number: params.last_round,
             utxos: vec![],
-            sequence: params.last_round,
+            sequence: params.last_round as u64,
             chain_id: params.genesis_id,
             is_destination_address_exist: true,
         })
